Sincronizar repositório com retrocompatibilidade
